@@ -48,12 +48,12 @@ class GLOP(ConicSolver):
         """Imports the solver."""
         import google.protobuf  # noqa F401
         import ortools  # noqa F401
-        if Version(ortools.__version__) < Version('9.4.0'):
+        if Version(ortools.__version__) < Version('9.5.0'):
             raise RuntimeError(f'Version of ortools ({ortools.__version__}) '
-                               f'is too old. Expected >= 9.4.0.')
-        if Version(ortools.__version__) >= Version('9.7.0'):
+                               f'is too old. Expected >= 9.5.0.')
+        if Version(ortools.__version__) >= Version('9.8.0'):
             raise RuntimeError('Unrecognized new version of ortools '
-                               f'({ortools.__version__}). Expected < 9.7.0.'
+                               f'({ortools.__version__}). Expected < 9.8.0. '
                                'Please open a feature request on cvxpy to '
                                'enable support for this version.')
 
